ARROW-15282: [C++][FlightRPC] Split out more gRPC-specific code
