+++ cpp/src/arrow/flight/serialization_internal.h
@@ -20,51 +20,12 @@
 
 #pragma once
 
-#include <memory>
-
 #include "arrow/flight/internal.h"
 #include "arrow/flight/transport.h"
-#include "arrow/flight/types.h"
-#include "arrow/ipc/message.h"
-#include "arrow/result.h"
 
 namespace arrow {
-
-class Buffer;
-
 namespace flight {
 namespace internal {
-
-/// Write Flight message on gRPC stream with zero-copy optimizations.
-// Returns Invalid if the payload is ill-formed
-// Returns true if the payload was written, false if it was not
-// (likely due to disconnect or end-of-stream, e.g. via an
-// asynchronous cancellation)
-arrow::Result<bool> WritePayload(
-    const FlightPayload& payload,
-    grpc::ClientReaderWriter<pb::FlightData, pb::PutResult>* writer);
-arrow::Result<bool> WritePayload(
-    const FlightPayload& payload,
-    grpc::ClientReaderWriter<pb::FlightData, pb::FlightData>* writer);
-arrow::Result<bool> WritePayload(
-    const FlightPayload& payload,
-    grpc::ServerReaderWriter<pb::FlightData, pb::FlightData>* writer);
-arrow::Result<bool> WritePayload(const FlightPayload& payload,
-                                 grpc::ServerWriter<pb::FlightData>* writer);
-
-/// Read Flight message from gRPC stream with zero-copy optimizations.
-/// True is returned on success, false if stream ended.
-bool ReadPayload(grpc::ClientReader<pb::FlightData>* reader, FlightData* data);
-bool ReadPayload(grpc::ClientReaderWriter<pb::FlightData, pb::FlightData>* reader,
-                 FlightData* data);
-bool ReadPayload(grpc::ServerReaderWriter<pb::PutResult, pb::FlightData>* reader,
-                 FlightData* data);
-bool ReadPayload(grpc::ServerReaderWriter<pb::FlightData, pb::FlightData>* reader,
-                 FlightData* data);
-// Overload to make genericity easier in DoPutPayloadWriter
-bool ReadPayload(grpc::ClientReaderWriter<pb::FlightData, pb::PutResult>* reader,
-                 pb::PutResult* data);
-
 // We want to reuse RecordBatchStreamReader's implementation while
 // (1) Adapting it to the Flight message format
 // (2) Allowing pure-metadata messages before data is sent
