+++ cpp/src/arrow/flight/transport/grpc/protocol_grpc_internal.cc
@@ -14,13 +14,12 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 
-#include "arrow/flight/protocol_internal.h"
+#include "arrow/flight/transport/grpc/protocol_grpc_internal.h"
 
 // NOTE(wesm): Including .cc files in another .cc file would ordinarily be a
 // no-no. We have customized the serialization path for FlightData, which is
 // currently only possible through some pre-processor commands that need to be
 // included before either of these files is compiled. Because we don't want to
 // edit the generated C++ files, we include them here and do our gRPC
-// customizations in protocol-internal.h
+// customizations in protocol_grpc_internal.h
 #include "arrow/flight/Flight.grpc.pb.cc"  // NOLINT
-#include "arrow/flight/Flight.pb.cc"       // NOLINT
